fix: restore author field to package.json to resolve cocoapods error
I had not realized that `pod install` would fail if the `authors` field
in the podspec was not provided. It resulted in the error:
>  ERROR | attributes: Missing required attribute `authors`.

Since the podspec pulls that field from package.json, I decided to restore
the author field there. I also added the contributors field to preserve
credit to the original author for their contributions.