fix: use correct method getFormat for barcode format
This was my mistake. I broke this when refactoring my previous commit
which introduced this change. I should have caught it in testing but I
made a mistake in the process.

I am going to invest in a better plan for testing changes to this module locally
before pushing to prevent pushing releases that don't work in the future.